+++ components/popup/Portal.tsx
@@ -2,14 +2,12 @@ import React, { PureComponent, CSSProperties, ReactPortal } from 'react';
 import ReactDOM from 'react-dom';
 import classnames from 'classnames';
 import Events from '../utils/events';
+import getMountNode from '../utils/getMountNode';
 import Mask from '../mask';
 import PropsType from './PropsType';
+import canUseDOM from '../utils/canUseDom';
 
 const IS_REACT_16 = !!ReactDOM.createPortal;
-
-function canUseDOM() {
-  return !!(typeof window !== 'undefined' && window.document && window.document.createElement);
-}
 
 export interface PortalProps extends PropsType {
   prefixCls?: string;
@@ -20,7 +18,7 @@ export interface PortalProps extends PropsType {
 export default class Portal extends PureComponent<PortalProps, any> {
   private enterTimer: number;
 
-  private parent: HTMLElement;
+  private mountNode: HTMLElement;
 
   private _container: HTMLDivElement;
 
@@ -69,25 +67,9 @@ export default class Portal extends PureComponent<PortalProps, any> {
 
     clearTimeout(this.enterTimer);
     if (this._container) {
-      this.parent.removeChild(this._container);
+      this.mountNode.removeChild(this._container);
     }
   }
-
-  getParent = () => {
-    const { getContainer } = this.props;
-    if (getContainer) {
-      if (typeof getContainer === 'function') {
-        return getContainer();
-      }
-      if (
-        typeof getContainer === 'object'
-        && getContainer instanceof HTMLElement
-      ) {
-        return getContainer;
-      }
-    }
-    return document.body;
-  };
 
   animationEnd = (e) => {
     if (e.target !== this.popup) {
@@ -144,7 +126,6 @@ export default class Portal extends PureComponent<PortalProps, any> {
   getComponent = () => {
     const {
       prefixCls,
-      className,
       animationType,
       animationDuration,
       direction,
@@ -157,7 +138,7 @@ export default class Portal extends PureComponent<PortalProps, any> {
     const animationState = visible ? 'enter' : 'leave';
 
     const cls = {
-      wrapper: classnames(`${prefixCls}__wrapper`, className, {
+      wrapper: classnames(`${prefixCls}__wrapper`, {
         [`za-fade-${animationState}`]: direction === 'center' && isPending,
       }),
       popup: classnames(prefixCls, {
@@ -249,7 +230,7 @@ export default class Portal extends PureComponent<PortalProps, any> {
   };
 
   renderPortal = (): ReactPortal | null => {
-    if (!canUseDOM()) {
+    if (!canUseDOM) {
       return null;
     }
     if (!IS_REACT_16) {
@@ -264,11 +245,12 @@ export default class Portal extends PureComponent<PortalProps, any> {
   };
 
   createContainer = () => {
+    const { className = '', prefixCls } = this.props;
     if (!this._container) {
       this._container = document.createElement('div');
-      this._container.className += 'popup-container';
-      this.parent = this.getParent();
-      this.parent.appendChild(this._container);
+      this._container.className += `${prefixCls}-container ${className}`;
+      this.mountNode = getMountNode(this.props);
+      this.mountNode.appendChild(this._container);
     }
     return this._container;
   };